Add shell control.
Now we can send commands to the spaceship from any shell.

$ echo "play" > ~/.spaceship/control

The primary motivation for this was to be able to call READ-DOCUMENT
from the shell so that all read documents are entered into the caching
and indexing system. See /spaceship/scripts/read-document.

I do envision not needing the shell at all at some point, but for now
it's still the best game in town.