Add LATEST-IRC.
LATEST-IRC will query the sentry and display all unread irc messages
in the *chats* emacs buffer.

I wrote this because I wanted to make it easier for me to follow the
'#lisp' irc channel. LATEST-IRC uniquely colors all handles and, in
addition, if the handle is listed in *WATCHING-HANDLES* the handle
will have a bright red background which makes it easy to pick out the
people that I wish to follow.