Add MAIL.
SPACESHIP> (mail "John" "Work on oneself" "Work on oneself must begin
with the driver. The driver is the mind. In order to be able to hear
the master's voice, the driver, first of all, must not sleep, that is,
he must wake up. Then it may prove that the master speaks a language
that the driver does not understand. The driver must learn this
language. When he has learned it, he will understand the master. But
concurrently with this he must learn to drive the horse, to harness it
to the carriage, to feed and to groom it, and to keep the carriage in
order.")