Add DEFUN-SIMPLE-MEMOIZED.
Defines a function with name NAME and single arguement ARG that
memoizes its return values in a synchronized hash table with test
TEST. If KEY is set to a function, then KEY is used on ARG to generate
the hash key to memoize. The NAME symbol MEMOIZE-HASH property is set
the hash for inspection. If TIMEOUT is set, use an EXPIRING-HASH-TABLE
with a timeout of TIMEOUT.