Add PING.
SPACESHIP> (ping "yahoo.com")
"64 bytes from ir1.fp.vip.sk1.yahoo.com (72.30.2.43): icmp_seq=1 ttl=46 time=113 ms"
"ir1.fp.vip.sk1.yahoo.com"
"72.30.2.43"
1
46
113

where I'm trying to go with this
================================

This coding might just seem like a drunken walk, but there really is a
goal in mind. What I am trying to do here is develop a system that
allows me to monitor all the machines and sources of information in
the world that I wish. A key part needed is the ability to move the
'sentry' over to another machine, get it running and then connect to
it. The problem I am facing is that SBCL executables are way too large
for this purpose at > 100M, so I have developed another project called
'probe' that compiles with ECL which produces amazingly small common
lisp executables at under 100K. These should be able to be easily
SCP'd over to a target machine. Note that I am not attempting to hack
any system that I do not already have SSH access, this is mainly an
extension of my current abilities.