Add a battery monitor.
This commit shows the culmination of a nice late night's
hacking. Agents, the paint server, and a musical note player. All so
that I don't have my darn thinkpad die in the middle of working in the
field. My last computer had no problem notifying my of imminent death,
this one just dies right in the middle of my flow. This commit should
fix that problem.